requirements.txt
moviepy==1.0.3
SpeechRecognition==3.9.0
openai>=1.0.0
Pillow==9.5.0
python-dotenv==0.21.1
requests==2.31.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openai-whisper==20231117
ffmpeg-python==0.2.0 